refactor: move issue actions into actions/issues/ subdirectory
Organize actions by domain: issues/ contains follow-up, execute,
close, check-status and their shared resolver. Keeps actions/
clean for future domain directories (e.g. projects/, external/).
