+++ apps/api/src/cron/actions/issues/follow-up.ts
@@ -1,7 +1,7 @@
 import { issueEngine } from '@/engines/issue'
 import { ensureWorking } from '@/routes/issues/_shared'
-import { resolveIssue } from './issue-resolver'
-import { registerAction } from './registry'
+import { registerAction } from '../registry'
+import { resolveIssue } from './resolver'
 
 registerAction('issue-follow-up', {
   description: 'Send a follow-up message to an issue',
