+++ apps/api/src/cron/actions/issues/index.ts
@@ -1,0 +1,5 @@
+// Import issue action modules to trigger self-registration
+import './follow-up'
+import './execute'
+import './close'
+import './check-status'
